Extract instance_request to avoid mutual imports
Also, update branch and fix tests for flake8 updates
